fix(48-03): clippy clean — workspace tests pass, zero warnings
- Fix clippy errors: transmute annotations, null-to-fn, unnecessary casts,
  missing_const_for_thread_local, result_unit_err, complex types
- Add type aliases NtCreateFileFn and SetFileInformationByHandleFn
- Replace transmute(null) with unwrap_or_else(panic) in trampolines
- Add #[allow(dead_code)] on DenyReturnValue, apply_deny_return, PipeError::Timeout
- Fix pcwstr_to_string to use ptr::add instead of offset with usize cast
- Verify: cargo test --workspace = 1798 passed, cargo clippy --workspace -D warnings = clean

+++ dlp-hook-dll/src/pipe_client.rs
@@ -26,6 +26,7 @@ thread_local! {
 
 /// Errors that can occur during pipe communication.
 #[derive(Debug, Clone, PartialEq)]
+#[allow(dead_code)]
 pub enum PipeError {
     /// The pipe server is not running or the pipe does not exist.
     ConnectionRefused,
@@ -141,7 +142,7 @@ fn connect_pipe(pipe_name: &str, timeout_ms: u32) -> Result<HANDLE, PipeError> {
         let handle = unsafe {
             CreateFileW(
                 windows::core::PCWSTR::from_raw(name_wide.as_ptr()),
-                (FILE_GENERIC_READ.0 | FILE_GENERIC_WRITE.0) as u32,
+                FILE_GENERIC_READ.0 | FILE_GENERIC_WRITE.0,
                 FILE_SHARE_NONE,
                 None,
                 OPEN_EXISTING,
